refactor: moderation bot
