+++ mini-app/package.json
@@ -8,12 +8,12 @@
     "dev": "pnpm run init:minio:local &&  dotenv -e ../.env -- cross-var node --inspect ./node_modules/next/dist/bin/next dev --port %MINI_APP_PORT%",
     "test-env": "dotenv -e ../.env -- cross-var  echo %MINI_APP_PORT%",
     "dev:local": "next dev",
-    "dev:local:moderation": "dotenv -e ../.env -- cross-var tsx  src/lib/initModerationBot.ts",
+    "dev:local:moderation": "dotenv -e ../.env -- cross-var tsx  src/moderationBot/startBot.ts",
     "dev:server": "dotenv -e .env -- cross-var node --inspect ./node_modules/next/dist/bin/next dev --port %MINI_APP_PORT%",
     "cron:test": "dotenv -e ../.env -- cross-var tsx src/cronJobs/runTests.ts",
     "start:local:poa": "dotenv -e ../.env -- cross-var tsx src/workers/poaWorker.ts",
     "start:poa": "tsx src/workers/poaWorker.ts",
-    "start:moderation": "tsx src/lib/initModerationBot.ts",
+    "start:moderation": "tsx src/src/moderationBot/startBot.ts",
     "start:local:socket": "dotenv -e ../.env -- cross-var nodemon --watch src/sockets --ext ts --exec tsx src/sockets/index.ts",
     "start:socket": "tsx src/sockets/index.ts",
     "test:socket:client": "dotenv -e ../.env -- cross-var nodemon --watch src/sockets --ext ts --exec tsx src/sockets/tests/test-client.ts",
