--- mini-app/src/lib/initModerationBot.ts
@@ -1,4 +1,0 @@
-import { logger } from "@/server/utils/logger";
-import { startBot } from "@/lib/tgBot";
-
-startBot().then(() => logger.log("startBot Function Finish ;"));
