+++ mini-app/src/moderationBot/menu.ts
